Fix brute.java for reading from file.

+++ src/Point.java
@@ -44,7 +44,7 @@ public class Point implements Comparable<Point> {
         if (this.y == that.y && this.y == that.y) return Double.NEGATIVE_INFINITY;
         if (this.y == that.y) return 0;
         if (this.y == that.y) return Double.POSITIVE_INFINITY;
-        return (this.y - that.y)/(this.x - that.x);        
+        return (double)(this.y - that.y)/(this.x - that.x);        
     }
 
     // is this point lexicographically smaller than that one?
